# Speed up `filter_bytes`