@alamb 2024-11-07
I also ran the benchmarks and see the same improvements 👍 

```
++ critcmp master speedup_filter
group                                                                         master                                 speedup_filter
-----                                                                         ------                                 --------------
filter context mixed string view (kept 1/2)                                   1.00    674.8±6.62µs        ? ?/sec    1.01    679.6±6.81µs        ? ?/sec
filter context mixed string view high selectivity (kept 1023/1024)            1.00   1196.4±2.54µs        ? ?/sec    1.02   1217.7±3.17µs        ? ?/sec
filter context mixed string view low selectivity (kept 1/1024)                1.01   1172.9±0.95ns        ? ?/sec    1.00   1164.3±2.62ns        ? ?/sec
filter context short string view (kept 1/2)                                   1.00    442.1±2.81µs        ? ?/sec    1.01    447.8±4.74µs        ? ?/sec
filter context short string view high selectivity (kept 1023/1024)            1.00    783.1±1.14µs        ? ?/sec    1.01    790.9±3.73µs        ? ?/sec
filter context short string view low selectivity (kept 1/1024)                1.00    936.5±1.46ns        ? ?/sec    1.01    947.4±5.04ns        ? ?/sec
filter context string (kept 1/2)                                              4.86      3.8±0.07ms        ? ?/sec    1.00   774.1±21.31µs        ? ?/sec
filter context string dictionary (kept 1/2)                                   1.00    244.2±1.43µs        ? ?/sec    1.00    243.6±1.24µs        ? ?/sec
filter context string dictionary high selectivity (kept 1023/1024)            1.00    248.9±1.50µs        ? ?/sec    1.00    248.1±0.92µs        ? ?/sec
filter context string dictionary low selectivity (kept 1/1024)                1.00    202.0±0.50µs        ? ?/sec    1.00    201.5±0.74µs        ? ?/sec
filter context string dictionary w NULLs (kept 1/2)                           1.00    341.8±0.75µs        ? ?/sec    1.00    342.5±0.93µs        ? ?/sec
filter context string dictionary w NULLs high selectivity (kept 1023/1024)    1.00    466.3±0.75µs        ? ?/sec    1.01    469.3±1.23µs        ? ?/sec
filter context string dictionary w NULLs low selectivity (kept 1/1024)        1.00    101.1±0.22µs        ? ?/sec    1.00    100.8±0.23µs        ? ?/sec
filter context string high selectivity (kept 1023/1024)                       1.63      2.5±0.06ms        ? ?/sec    1.00  1523.2±31.32µs        ? ?/sec
filter context string low selectivity (kept 1/1024)                           1.23   1993.4±3.93ns        ? ?/sec    1.00   1614.9±4.07ns        ? ?/sec
```