@alamb 2024-11-08
Results of rerunning

```
++ critcmp master speedup_filter
group                                                                         master                                 speedup_filter
-----                                                                         ------                                 --------------
filter context mixed string view (kept 1/2)                                   1.00    676.9±1.46µs        ? ?/sec    1.00    680.2±3.73µs        ? ?/sec
filter context mixed string view high selectivity (kept 1023/1024)            1.00   1195.0±3.18µs        ? ?/sec    1.02   1217.4±3.71µs        ? ?/sec
filter context mixed string view low selectivity (kept 1/1024)                1.01   1167.2±2.60ns        ? ?/sec    1.00   1156.4±1.15ns        ? ?/sec
filter context short string view (kept 1/2)                                   1.00    441.9±3.60µs        ? ?/sec    1.02    451.0±2.40µs        ? ?/sec
filter context short string view high selectivity (kept 1023/1024)            1.00    780.6±2.07µs        ? ?/sec    1.01    790.6±1.75µs        ? ?/sec
filter context short string view low selectivity (kept 1/1024)                1.01    930.2±2.89ns        ? ?/sec    1.00    918.1±4.38ns        ? ?/sec
filter context string (kept 1/2)                                              4.89      3.7±0.09ms        ? ?/sec    1.00   757.8±12.84µs        ? ?/sec
filter context string dictionary (kept 1/2)                                   1.02    247.1±1.64µs        ? ?/sec    1.00    242.5±0.56µs        ? ?/sec
filter context string dictionary high selectivity (kept 1023/1024)            1.02    252.3±1.74µs        ? ?/sec    1.00    248.4±1.13µs        ? ?/sec
filter context string dictionary low selectivity (kept 1/1024)                1.00    201.8±0.46µs        ? ?/sec    1.00    200.9±0.48µs        ? ?/sec
filter context string dictionary w NULLs (kept 1/2)                           1.00    341.7±0.77µs        ? ?/sec    1.00    341.9±0.52µs        ? ?/sec
filter context string dictionary w NULLs high selectivity (kept 1023/1024)    1.00    466.6±1.22µs        ? ?/sec    1.01    469.1±3.93µs        ? ?/sec
filter context string dictionary w NULLs low selectivity (kept 1/1024)        1.00    101.0±0.09µs        ? ?/sec    1.00    101.0±0.38µs        ? ?/sec
filter context string high selectivity (kept 1023/1024)                       1.67      2.5±0.07ms        ? ?/sec    1.00  1491.5±32.49µs        ? ?/sec
filter context string low selectivity (kept 1/1024)                           1.23   1989.6±5.27ns        ? ?/sec    1.00   1622.9±4.10ns        ? ?/sec

```