@Dandandan 2024-11-08
Interesting, so it seems to have even bigger effect on other machines - my result is from M1 pro